third_party/WebKit/LayoutTests/http/tests/appcache/modified-manifest.html

<html manifest="resources/modified-manifest.php">
<script src="/js-test-resources/js-test.js"></script>
<script>
var jsTestIsAsync = true;
description("Test that a manifest that changes during an update fails.");

var gotCheckingEvent = false;
function onChecking() { gotCheckingEvent = true; }


var gotDownloadingEvent = false;
function onDownloading() { gotDownloadingEvent = true; }

function onError(e)
{
    shouldBeTrue("gotCheckingEvent");
    shouldBeTrue("gotDownloadingEvent");
    shouldBe("window.applicationCache.status", "applicationCache.UNCACHED");

    event = e;
    shouldBeEqualToString("event.reason", "changed");
    shouldBeEqualToString("event.url", "");
    shouldBe("event.status", "0");
    shouldBeTrue("'message' in event");

    finishJSTest();
}

applicationCache.addEventListener('checking', onChecking, false);
applicationCache.addEventListener('downloading', onDownloading, false);
applicationCache.addEventListener('error', onError, false);

</script>
</html>
