third_party/WebKit/LayoutTests/http/tests/appcache/modified-manifest-expected.txt

Test that a manifest that changes during an update fails.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gotCheckingEvent is true
PASS gotDownloadingEvent is true
PASS window.applicationCache.status is applicationCache.UNCACHED
PASS event.reason is "changed"
PASS event.url is ""
PASS event.status is 0
PASS 'message' in event is true
PASS successfullyParsed is true

TEST COMPLETE
